% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc.R
\name{bar}
\alias{bar}
\title{Lorem ipsum dolor sit amet, consectetuer adipiscing elit. Aenean commodo ligu}
\usage{
bar(...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[=foo]{foo}}
  \describe{
    \item{\code{a}}{Maecenas et nisi vitae purus rutrum vulputate vitae a orci. Sed leo.}
    \item{\code{b}}{Donec sed mollis odio. Donec lacinia fermentum lectus, nec posuere.}
    \item{\code{c}}{Vivamus ultricies eget turpis vel molestie. Vestibulum dapibus nam.}
  }}
}
\description{
Lorem ipsum dolor sit amet, consectetuer adipiscing elit. Aenean commodo ligu
}
