% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc.R
\name{foo}
\alias{foo}
\title{Quisque porttitor mattis lorem ac commodo. Donec volutpat in mi sed porttitor}
\usage{
foo(a, b, c)
}
\arguments{
\item{a}{Maecenas et nisi vitae purus rutrum vulputate vitae a orci. Sed leo.}

\item{b}{Donec sed mollis odio. Donec lacinia fermentum lectus, nec posuere.}

\item{c}{Vivamus ultricies eget turpis vel molestie. Vestibulum dapibus nam.}
}
\description{
Quisque porttitor mattis lorem ac commodo. Donec volutpat in mi sed porttitor
}
